Minor fixes and READMEs pending
