+++ samples/python/81.skills-skilldialog/dialog-skill-bot/authentication/allow_callers_claims_validation.py
@@ -19,13 +19,18 @@ class AllowedCallersClaimsValidator:
         # that consists of the list of parent bot ids that are allowed to access the skill
         # to add a new parent bot simply go to the AllowedCallers and add
         # the parent bot's microsoft app id to the list
-        self._allowed_callers = config.ALLOWED_CALLERS
+        caller_list = getattr(config, self.config_key)
+        if caller_list is None:
+            raise TypeError(
+                f"\"{self.config_key}\" not found in configuration."
+            )
+        self._allowed_callers = caller_list
 
     @property
     def claims_validator(self) -> Callable[[List[Dict]], Awaitable]:
         async def allow_callers_claims_validator(claims: Dict[str, object]):
-            # if _allowedCallers is None we allow all calls
-            if self._allowed_callers and SkillValidation.is_skill_claim(claims):
+            # if allowed_callers is None we allow all calls
+            if "*" not in self._allowed_callers and SkillValidation.is_skill_claim(claims):
                 # Check that the appId claim in the skill request is in the list of skills configured for this bot.
                 app_id = JwtTokenValidation.get_app_id_from_claims(claims)
                 if app_id not in self._allowed_callers:
