+++ samples/python/81.skills-skilldialog/dialog-skill-bot/bots/__init__.py
@@ -2,6 +2,5 @@
 # Licensed under the MIT License.
 
 from .skill_bot import SkillBot
-from .activity_router_dialog import ActivityRouterDialog
 
-__all__ = ["SkillBot", "ActivityRouterDialog"]
+__all__ = ["SkillBot"]
