+++ samples/python/81.skills-skilldialog/dialog-skill-bot/dialogs/__init__.py
@@ -1,18 +1,18 @@
+from .activity_router_dialog import ActivityRouterDialog
 from .booking_details import BookingDetails
 from .booking_dialog import BookingDialog
 from .cancel_and_help_dialog import CancelAndHelpDialog
 from .date_resolver_dialog import DateResolverDialog
 from .dialog_skill_bot_recognizer import DialogSkillBotRecognizer
 from .location import Location
-from .oauth_test_dialog import OAuthTestDialog
 
 
 __all__ = [
+    "ActivityRouterDialog",
     "BookingDetails",
     "BookingDialog",
     "CancelAndHelpDialog",
     "DateResolverDialog",
     "DialogSkillBotRecognizer",
     "Location",
-    "OAuthTestDialog",
 ]
